refactor(reporter): split monolithic file and improve SRP compliance
